factor DFT external function into separate file
To be linked into the CASPER app, which does not include sarbp.cpp.
